+++ Plateformer-2D-Rubika/Assets/Scripts/CameraControllers/CameraControllerMain.cs
@@ -10,7 +10,7 @@ public class CameraControllerMain : MonoBehaviour
 
     [Header("Settings")]
     [SerializeField] bool followPlayer;
-    [SerializeField] int orthographicSize;
+    [SerializeField] float orthographicSize;
     [SerializeField] Vector2 damping;
     [SerializeField] Vector2 deadZone;
